+++ services/issues-service/src/bootstrap/data-source.ts
@@ -3,7 +3,7 @@ import { env } from "@/env";
 
 export const dataSource = new DataSource({
   type: "postgres",
-  url: env.ISSUE_TRACKER_POSTGRES_CLUSTER_URL,
+  url: env.ISSUES_DATABASE_URL,
   entities: ["src/entities/*.{ts,js}"],
   synchronize: true,
 });